% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2013.csv}
\alias{accident_2013.csv}
\title{Data are from the US National Highway Traffic Safety Administration's Fatality
Analysis Reporting System (FARS), which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered in motor vehicle
traffic crashes.  For the purposes of this function, only data for the years
2013, 2014 and 2015 are considered (data included with the package).}
\format{
Dataset with 30,202 rows and 50 variables for year 2013.
}
\source{
\url{https://www-fars.nhtsa.dot.gov/Main/index.aspx}
}
\usage{
accident_2013.csv
}
\description{
Data are from the US National Highway Traffic Safety Administration's Fatality
Analysis Reporting System (FARS), which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered in motor vehicle
traffic crashes.  For the purposes of this function, only data for the years
2013, 2014 and 2015 are considered (data included with the package).
}
\keyword{datasets}
