% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of a file containing the annual data for a specific
year of the survey. The file is assumed to be in comma separated values (csv)
format.}
}
\value{
This function returns the data contained in the file as a tibble.  A
tibble is a special version of a data frame designed to work with the "tidyverse"
set of R packages.  If the file cannot be found, an error message "file ...
does not exist" is generated.
}
\description{
This is a function that reads in This function reads data from .csv file
from the US ational Highway Traffic Safety Administration's Fatality Analysis
Reporting System (FARS), which is a nationwide census that provides yearly data
regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
