% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A data set covering the indicated years, loaded into the R environment.
}
\description{
This function reads in a vector of the years to be examined in the Fatality
Analysis Reporting System (FARS) data and tests whether the information for the
specific year chosen is available.  If not, the function informs the user that
the year is invalid. Ancillary function used by \code{fars_summarize_years}.
}
\examples{
fars_read_years(2013, 2014, 2015)

}
