% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with a list of years to be summarized.}
}
\value{
A data set summarizing accidents by year and month, with other relevant
information.
}
\description{
This function summarizes accident data, by year and month, for the years input
by the user.
}
\examples{
fars_summarize_years(c(2014, 2015))

}
