% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which the data was collected.}
}
\value{
The function returns a character vector (the filename) containing a
combination of text and the year passed to the function.
}
\description{
This is a function that takes a given year as a string or integer (e.g. 2015)
and inserts into the name of a file that saves Fatality Analysis Reporting System
(FARS) information.
}
\examples{
make_filename(2015)

}
