% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the points of the accidents reported in the data set by location}
\usage{
fars_map_state(state.num, year)
}
\description{
This function plots a point at the location where the accidents reported for each year
happened. You must provide the year and the state where you want to plot the accidents.
It may be the case that there were no accidents reported in a state for a given year. 
In that case there will not be any points to plot. Also, the number rovided for the
state must be in the desrired range, otherwise you'll get an error if the number given
to the function is not associated with a particuar state.
}
\details{
@param state.num an integer indicating the id for the state to be plotted
 @inheritParams make_filename
 @return a map plotting the points at the location of the accdents reported for the 
 given state
 @importFrom dyplyr filter
 @importFrom maps map
 @importFrom graphics points
 
 @examples 
 fars_map_state(01,2014)
 
 @export
}
