% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map of fatalities}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that is an identifier for a certain state.}

\item{year}{A character string or integer representing the year of the
FARS data.}
}
\value{
This function will print a map to the graphics device and will
 return null
}
\description{
This function will create a map showing a single state and all the fatalies
for a given year. Each fatality will show up as a dot on a map.
}
\examples{
\dontrun{
# map fatalities in Florida for 2014 and then for 2015
fars_map_state(12,"2014")
fars_map_state(12, 2015)

# Using an invalid year will result in an error ...
fars_map_state(12,"1892")
# Error in fars_read(filename) :
# file 'accident_1892.csv.bz2' does not exist

# Using an invalid state number will result in an error ...
fars_map_state(200,"2014")
# Error in fars_map_state(200, "2014") : invalid STATE number: 200
}

}
