% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map displaying location of fatal injuries for selected state and year}
\source{
Data source: US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS),
        which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a interger or string indicating state code, please see notes.}

\item{year}{a interger or string indicating year.}
}
\value{
A map showing location of fatal injuries for selected state and year.
        If the state code (\code{state.num}) does not exist, return error message "invalid STATE number".
        If the file for given year does not exsit, return error message "file ... does not exist".
}
\description{
This function will display a map of fatal injuries location for selected state and year.
}
\note{
State code:
  01: Alabama,
  02: Alaska,
  04: Arizona,
  05: Arkansas,
  06: California,
  08: Colorado,
  09: Connecticut,
  10: Delaware,
  11: District of Columbia,
  12: Florida,
  13: Georgia,
  15: Hawaii,
  16: Idaho,
  17: Illinois,
  18: Indiana,
  19: Iowa,
  20: Kansas,
  21: Kentucky,
  22: Louisiana,
  23: Maine,
  24: Maryland,
  25: Massachusetts,
  26: Michigan,
  27: Minnesota,
  28: Mississippi,
  29: Missouri,
  30: Montana,
  31: Nebraska,
  32: Nevada,
  33: New Hampshire,
  34: New Jersey,
  35: New Mexico,
  36: New York,
  37: North Carolina,
  38: North Dakota,
  39: Ohio,
  40: Oklahoma,
  41: Oregon,
  42: Pennsylvania,
  43: Puerto Rico,
  44: Rhode Island,
  45: South Carolina,
  46: South Dakota,
  47: Tennessee,
  48: Texas,
  49: Utah,
  50: Vermont,
  51: Virginia,
  52: Virgin Islands,
  53: Washington,
  54: West Virginia,
  55: Wisconsin,
  56: Wyoming.
}
\examples{
\dontrun{fars_map_state(40, 2013)}

}
\seealso{
\code{\link{fars_summarize_years}}

\code{\link{make_filename}}

\code{\link{fars_read}}
}
