% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Makes a geographical map of FARS incidents for given state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A state number}

\item{year}{An year}
}
\description{
Loads FARS data for specified year and plots incidents happened in given state on a map
using incident's geographical position.

If a wrong state number is provided, the execution is stopped.
If a file for given year does not exist or cannot be loaded, a warning is logged.
}
\examples{
fars_map_state(1, 2014)

}
\seealso{
\code{\link{fars_summarize_years}}
}
