% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer the provides the number of the state whose data
is to be retrieved.  See (link) for a listing of state numbers.}

\item{year}{An integer which contains the year of interest.}
}
\value{
A plot with accident locations plotted on a map of the selected state.
}
\description{
Displays a plot with a state map including the accidents location by year
If the state.num is invalid, the function shows an error.
}
\examples{
fars_map_state(25, 2013)
#'
}
