% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a dataset into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{- provides the name of the file with the data to be read into R}
}
\value{
a table data frame in R
}
\description{
This function read an existing file in your directory and convert it into a
data table in R. Keep in mind that the file to be imported into R must be in
your directory, otherwise you may get an error.
}
\examples{
fars_read(accident_2013)

}
