% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read the FARS file}
\source{
Data source: US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS),
        which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
        For more information, please see \url{https://crashstats.nhtsa.dot.gov/#/DocumentTypeList/23}.
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string indicates the name of the file. Can use \code{make_filename} to generate a filename.}
}
\value{
This function returns a dataframe if the selected file exist;
        return an error message "file ... does not exist" if the file not exist.
}
\description{
This function will search for the input filename in the current directory, if the file exist,
             it will read in the file (.csv format) and store the file to current directory;
             or return error if the file does not exist.
}
\note{
To generate a file name, use \code{make_filename}.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
\seealso{
\code{\link{make_filename}}
}
