% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A name of file to load FARS data from. File must be in CSV format.}
}
\value{
A data frame containing FARS data
}
\description{
Loads FARS data from provided file in CSV format.

If a file does not exist, an execution is stopped.
}
\examples{
df <- fars_read(make_filename(2015))

}
\seealso{
\code{\link{make_filename}}, \code{\link{fars_read_years}}
}
