% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads datafiles into R at once}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector containing the years for the files to be read into R. Keep in 
mind that you must provide the years of data that are available as datasets, otherwise
you'll get an error.}
}
\value{
a dataframe(s) in R
}
\description{
This function reads one or more datafiles into R at once.
}
\examples{
fars_read_years(c(2014,2015))

}
