% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read one or more FARS data files.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character string or vector of character strings representing 
the year(s) of the FARS data. An integer or vector of integers may also
be used for input; which represent the year(s) of the FARS data.}
}
\value{
This function returns a list containing all of the FARS data files
that were read. Each element of the list will be a tbl_df object containing
the FARS data corresponding to the order in which its year appeared in the
input parameter. Each tbl_df object will contain only two columns as follows:
'Month' and 'year', corresponding to the FARS data object it represents.
Each row of each tbl_df object represents one fatality that occured in the
month noted and the year noted.
}
\description{
This function will read one or more FARS data files designated by the input
years which can be a single year or a vector of years.
}
\examples{
# using character strings ...
farsYearsListA <- fars_read_years("2014")
farsYearsListB <- fars_read_years(c("2013","2015"))

# using integers  ...
farsYearsListA <- fars_read_years(2014)
farsYearsListB <- fars_read_years(c(2013,2015))

# using invalid year for the input will result in an error ...
farsYearsListBad <- fars_read_years("1894")
#  Warning message:
#  In value[[3L]](cond) : invalid year: 1894

}
