% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the FARS data with given year}
\source{
Data source: US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS),
        which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector with list of years.}
}
\value{
If the given year exist, return a dataframe with the given year and month of the fatal injuries incidents (long format);
        if the file does not exist, return NULL with warning message "invalid year: ..."
}
\description{
This function will search and read the FARS files in the current directory with the given years, then
             return dataframe with the given years and month of the fatal injuries incidents (long format, each year is a seperate dataframe),
             or NULL if the file not exist.
}
\examples{
\dontrun{fars_read_years(2013)}
\dontrun{fars_read_years(c(2013, 2014))}

}
\seealso{
\code{\link{fars_read}}

\code{\link{make_filename}}
}
