% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector of years to load data for}
}
\value{
A list containing FARS data for given years
}
\description{
Loads FARS data for multiple years. Data files must be in current working directory.

If a file for given year does not exist or cannot be loaded, a warning is logged.
}
\examples{
df_list <- fars_read_years(c(2013, 2014))

}
\seealso{
\code{\link{fars_read}}
}
