% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Provides the sample size for datasets}
\usage{
fars_summarize_years(years)
}
\description{
This function gives the sample size for each dataset imputed into it.
}
\details{
@param years
 @inheritParams fars_read_years
 @importFrom dyplyr bind_rows 
 @importFrom dyplyr group_by
 @importFrom dyplyr summarize
 @importFrom tidyr spread
 @return a vector containing the sample size by dataset (year)
 
 @examples 
 fars_summarize_years(c(2013,2014,2015))
 
 @export
}
