% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by year}
\source{
Data source: US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS),
        which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector with list of years.}
}
\value{
A summarize table with the given year and number of injuries counts per month (in wide format);
        if the data with given year does not exist, return NULL with warning message "invalid year: (given_year)".
}
\description{
This function will summarize the number of injuries counts per month for the given years,
             or return warning if the data with given year does not exist.
}
\examples{
fars_summarize_years(2013), fars_summarize_years(c(2013, 2014))

}
