% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Provides a number of FARS incidents grouped by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector of years to load data for}
}
\value{
A data frame containing number of incidents per year and month
}
\description{
Loads FARS data for multiple years and groups them by year and month.

If a file for given year does not exist or cannot be loaded, a warning is logged.
}
\examples{
df_grp <- fars_summarize_years(c(2013, 2014))

}
\seealso{
\code{\link{fars_read}}, \code{\link{fars_read_years}}
}
