% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates the name for data file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year of the data to be read as integer. Keep in mind that if not given
an integer as argument, you may get an error.}
}
\value{
this function returns a string with the name of the file "accident" and the 
year provided as imput.
}
\description{
This function creates the name for the data file to be read into R.
}
\examples{
make_filename(2014)

}
