% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a correctly formatted FARS file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string representing the year of the FARS data.
An integer value representing a year may also be used as input.}
}
\value{
This function returns a string representing the correctly
  formatted filename for the FARS data file.
}
\description{
This function will make a correctly formatted FARS file name using the
input that represents the year of the FARS data.
}
\examples{
# Using charactger string input ...
filename2015 <- make_filename("2015")
#  "accident_2015.csv.bz2"

# Using integer input ...
filename2017 <- make_filename(2017)

}
