% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer or string indicating the input year.}
}
\value{
This function will return the filname as a string format using the input year: accident_(year).csv.bz2"
}
\description{
This function will create a .csv filename with input year.
}
\note{
This function will only create a filename, to read in the file, use \code{fars_read}.
}
\examples{
make_filename(2013) ## output "accident_2013.csv.bz2"

}
\seealso{
\code{\link{fars_read}}
}
