% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Make a name of file containing FARS data for given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year to load data for}
}
\value{
A file name (without PATH) to be used with fars_read function
}
\description{
Constructs a file name of FARS data file for specified year
}
\examples{
df <- fars_read(make_filename(2015))

}
\seealso{
\code{\link{fars_read}}
}
